+++ R/dbi-connection.R
@@ -220,7 +220,7 @@ setMethod("dbListFields", c("OdbcConnection", "character"),
 setMethod("dbGetInfo", "OdbcConnection",
   function(dbObj, ...) {
     info <- connection_info(dbObj@ptr)
-    structure(info, class = c(info$dbms.name, "driver_info", "list"))
+    structure(info, class = c(info$dbms.name, "list"))
   }
 )