+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Use correct parent class for Oracle (#685).
 
+* SQL Server: correctly enumerate schemas across databases in connections pane
+  (@detule, #527)
+
 # odbc 1.4.0
 
 ## Major changes