+++ NEWS.md
@@ -3,6 +3,9 @@
 * Added a function `odbcListConfig()` to help locate configuration files on
   macOS and Linux (@simonpcouch, #565).
 
+* Use correct parent class for Oracle (#685).
+
+
 # odbc 1.4.0
 
 ## Major changes