+++ NEWS.md
@@ -1,6 +1,10 @@
 # odbc (development version)
 
+* Added a function `odbcListConfig()` to help locate configuration files on
+  macOS and Linux (@simonpcouch, #565).
+
 * Use correct parent class for Oracle (#685).
+
 
 # odbc 1.4.0